Auto-detect file type from content
Auto-detecting the file type based on the first few characters
is rather easy. Ipynb files start with an opening curly brace,
python files are expected to start with '#!/usr/bin/env python3'.
By checking the first few characters of the file to detect the
file type, the reliance on the filename extension can be dropped.
This allows to read files from stdin, if desired.